+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # WebTransport APIs tests changelog
 
-## Test Go WebTransport and WebSockets
+## Test Go WebTransport and WebSocket
 
-- Implemented a WebSockets client-server interaction with gorilla/websocket library for Go environment.
+- Implemented a WebSocket client-server communication with gorilla/websocket library for Go environment.
 - Implemented performance test for Go WebTransport client-server application.
 - Implemented performance test for Go WebSockets client-server application.
+
+## Test Browser WebTransport and WebSocket
+
+- Implemented a WebSocket client-server with websockets/ws library for nodejs environment.
+- Implemented performance test for browser WebTransport client-server application.
+- Implemented performance test for browser WebSockets client-server application.